+++ controllers/adminController.js
@@ -1,0 +1,40 @@
+const adminController = {
+    index: (req, res, next) => {
+        let tabActive = {homeAct: "inactive",
+                        adminAct: "active",
+                        financeiroAct: "inactive",
+                        clientesAct: "inactive",
+                        funcionariosAct: "inactive",
+                        pdvAct: "inactive"};
+        res.render('admin/admin', { title: 'Express', tabs: tabActive });
+    },
+    indexProdutos: (req, res, next) => {
+        let tabActive = {homeAct: "inactive",
+                        adminAct: "active",
+                        financeiroAct: "inactive",
+                        clientesAct: "inactive",
+                        funcionariosAct: "inactive",
+                        pdvAct: "inactive"};
+        res.render('admin/produtos', { title: 'Express', tabs: tabActive });
+    },
+    indexEstoque: (req, res, next) => {
+        let tabActive = {homeAct: "inactive",
+                        adminAct: "active",
+                        financeiroAct: "inactive",
+                        clientesAct: "inactive",
+                        funcionariosAct: "inactive",
+                        pdvAct: "inactive"};
+        res.render('admin/estoque', { title: 'Express', tabs: tabActive });
+    },
+    indexReceitas: (req, res, next) => {
+        let tabActive = {homeAct: "inactive",
+                        adminAct: "active",
+                        financeiroAct: "inactive",
+                        clientesAct: "inactive",
+                        funcionariosAct: "inactive",
+                        pdvAct: "inactive"};
+        res.render('admin/receitas', { title: 'Express', tabs: tabActive });
+    }
+};
+
+module.exports = adminController;
