+++ database/produtos.json
@@ -1,0 +1,7 @@
+[{
+    "nome":"Brigadeiro",
+    "descrição":"Brigadeiro de cholocate",
+    "quantidade":10,
+    "valor":10.99,
+    "imagem":"/images/produtos/brigadeiro.png"
+    }]
