+++ docs/syshome.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Candy Sphere</title>
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="/css/style-system.css">
+    <link rel="stylesheet" href="css/style-system.css">
 </head>
 <body>
     <header>
