+++ views/admin/estoque.ejs
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+
+    <%- include('../partials/head') %>
+
+    <body>
+
+        <%- include('../partials/header') %>
+
+        <main id="administracao">
+
+        </main>
+
+        <%- include('../partials/footer') %>
+
+        <%- include('../partials/scripts') %>
+
+    </body>
+
+</html>
